jira2slack/requirements.txt
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
feedparser==5.2.1
idna==2.9
PyYAML==5.3.1
requests==2.23.0
slackweb==1.0.5
soupsieve==2.0
urllib3==1.25.9
